Доработка сброса пароля версия 2
