+++ src/main/resources/templates/main-page.html
@@ -41,7 +41,7 @@
                                 </label>
 
 
-                                <a style="font-size: 10pt" href="#" id="open-modal-3" data-dismiss="modal">Забыли пароль?</a>
+                                <a style="font-size: 10pt" href="#" id="open-modal-3" th:text="#{message.resetPassword}" data-dismiss="modal">Забыли пароль?</a>
 
                         </div><br>
                         <div class="form-item-button">
@@ -93,7 +93,7 @@
         <!-- Modal content-->
         <div class="modal-content">
             <div class="modal-header">
-                <h4 class="modal-title">Регистрация</h4>
+                <h4 class="modal-title" id="message-reset-password">Регистрация</h4>
                 <button type="button" class="close" data-dismiss="modal">×</button>
 
             </div>
@@ -148,7 +148,7 @@
 
                     <div class="form-item-login">
                         <label class="input-login">
-                            Введите EMAIL: <input name="email" placeholder="email" type="text" required>
+                            Введите EMAIL: <input name="email" placeholder="email" type="email" required>
                         </label>
                     </div>
                     <div class="form-reg-button">
